server/requirements.txt
asgiref==3.9.1
coverage==7.5.4
Django==5.0.14
django-cors-headers==4.3.1
djangorestframework==3.15.2
drf-yasg==1.21.10
gunicorn==23.0.0
inflection==0.5.1
packaging==25.0
pillow==10.3.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.2.0
